Added the drink removal function. (#44)
This required to rework the drinker data hosting, now in a singleton
repository (not great, to improve later, see #42).

A temporary implementation used a view model, but it was removed as of
little use and redundant. The dependencies were kept though (see #43)

The DrinkerFragment is now updated with the previous (or default for now) values, coming from the DrinkerRepository. The retrieval of values from previous session is still to do.

Fixed #38